Refactor nav links
