--- config/navLinks.ts
@@ -1,13 +1,0 @@
-type NavLink = {
-  title: string;
-  href: string;
-};
-
-export const navLinks: NavLink[] = [
-  { title: "Home", href: "/" },
-  // { title: "Blog", href: "/blog" },
-  { title: "About", href: "/about" },
-  { title: "Journals", href: "/journals" },
-  { title: "Projects", href: "/projects" },
-  // { title: "Resume", href: "/resume" },
-];
